+++ aiida/src/main/java/energy/eddie/aiida/models/permission/dataneed/AiidaLocalDataNeed.java
@@ -8,6 +8,7 @@ import energy.eddie.api.agnostic.aiida.AiidaAsset;
 import energy.eddie.api.agnostic.aiida.AiidaSchema;
 import energy.eddie.api.agnostic.aiida.ObisCode;
 import energy.eddie.api.agnostic.aiida.ObisCodeConverter;
+import energy.eddie.cim.agnostic.PermissionCommand;
 import energy.eddie.dataneeds.needs.aiida.AiidaDataNeed;
 import energy.eddie.dataneeds.needs.aiida.AiidaDataNeedInterface;
 import energy.eddie.dataneeds.utils.cron.CronExpressionConverter;
@@ -63,6 +64,14 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
     protected boolean acknowledgementRequired;
 
     @ElementCollection(fetch = FetchType.EAGER)
+    @CollectionTable(name = "aiida_local_data_need_permission_commands",
+            joinColumns = @JoinColumn(name = "data_need_id"))
+    @Column(name = "permission_command")
+    @Enumerated(EnumType.STRING)
+    @JsonProperty
+    protected Set<PermissionCommand.Action> allowedPermissionCommands;
+
+    @ElementCollection(fetch = FetchType.EAGER)
     @CollectionTable(name = "aiida_local_data_need_schemas", joinColumns = {@JoinColumn(name = "data_need_id", referencedColumnName = "data_need_id")})
     @Column(name = "schema")
     @JsonProperty
@@ -99,6 +108,7 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
         this.asset = dataNeed.asset();
         this.dataTags = Objects.requireNonNullElse(dataNeed.dataTags(), Set.of());
         this.acknowledgementRequired = dataNeed.acknowledgementRequired();
+        this.allowedPermissionCommands = Objects.requireNonNullElse(dataNeed.allowedPermissionCommands(), Set.of());
     }
 
     public String name() {
@@ -133,6 +143,11 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
     @Override
     public boolean acknowledgementRequired() {
         return acknowledgementRequired;
+    }
+
+    @Override
+    public Set<PermissionCommand.Action> allowedPermissionCommands() {
+        return allowedPermissionCommands;
     }
 
     @Override